+++ uninstall.php
@@ -19,3 +19,4 @@ delete_option( 'darkadmin_allowed_users' );
 delete_option( 'darkadmin_user_access_mode' );
 delete_option( 'darkadmin_preset' );
 delete_option( 'darkadmin_excluded_pages' );
+delete_option( 'darkadmin_db_version' );